third_party/blink/web_tests/platform/generic/fast/lists/ol-nested-list-dynamic-insert-expected.txt

This tests that list item markers do not change when adding new items to a nested list.

1 One
2 Two
3 Three
4 Four
5 Five One Twenty five Twenty six
-> -> 1 One
-> -> 25 Twenty five
-> -> 26 Twenty six
6 Six
7 Seven
